requirements.txt
nextcord
aiohttp
fastapi
pydantic
uvicorn
pyyaml